+++ client/src/App.css
@@ -71,3 +71,7 @@ input::-webkit-inner-spin-button {
   padding-right: 10px;
   margin-top: 10px;
 }
+
+.navbarImg {
+  width: 50px;
+}
