+++ client/src/components/AddWorkout.js
@@ -1,18 +1,8 @@
 import React, { useState } from "react";
-import Grid from "@material-ui/core/Grid";
-import TextField from "@material-ui/core/TextField";
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-import FormControl from "@material-ui/core/FormControl";
-import FormLabel from "@material-ui/core/FormLabel";
-import RadioGroup from "@material-ui/core/RadioGroup";
-import Radio from "@material-ui/core/Radio";
-import Select from "@material-ui/core/Select";
-import MenuItem from "@material-ui/core/MenuItem";
-import Slider from "@material-ui/core/Slider";
-import Button from "@material-ui/core/Button";
+import {FormControl, Button, Input} from "@mui/material";
 import { FormGroup, InputLabel } from "@mui/material";
 
-export const AddWorkout = () => {
+export default function AddWorkout () {
   return (
     <div>
       <h1>Add Workout</h1>
